+++ resourcebundle-inspector/src/test/java/de/alpharogroup/resourcebundle/properties/PropertiesFileExtensionsTest.java
@@ -42,7 +42,7 @@ import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
 import de.alpharogroup.collections.set.SetFactory;
-import de.alpharogroup.file.FileExtension;
+import de.alpharogroup.io.file.FileExtension;
 import de.alpharogroup.file.search.PathFinder;
 import de.alpharogroup.lang.ClassExtensions;
 import de.alpharogroup.lang.PackageExtensions;
@@ -75,16 +75,6 @@ public class PropertiesFileExtensionsTest
 	public void testGetProjectName() throws IOException
 	{
 		final String projectName = PropertiesFileExtensions.getProjectName();
-		assertNotNull(projectName);
-	}
-
-	/**
-	 * Test method for {@link PropertiesFileExtensions#getProjectNameQuietly(String)}.
-	 */
-	@Test
-	public void testGetProjectNameQuietly()
-	{
-		final String projectName = PropertiesFileExtensions.getProjectNameQuietly("foo");
 		assertNotNull(projectName);
 	}